move source code to folders
Cleaner and makes life easier when browsing tree from github
